+++ spec/nocode/steps/serialize/csv_spec.rb
@@ -9,11 +9,30 @@ describe Nocode::Steps::Serialize::Csv do
   let(:context)   { Nocode::Context.new(registers: registers) }
   let(:options)   { { 'register' => 'a' } }
 
-  it 'copies register value to another register' do
+  it 'serializes register contents as CSV' do
     step.perform
 
     expected = "id,name\n1,bozo\n"
 
     expect(context.register('a')).to eq(expected)
+  end
+
+  context 'when contents is an array of arrays' do
+    let(:registers) do
+      {
+        'a' => [
+          %w[id name],
+          %w[1 bozo]
+        ]
+      }
+    end
+
+    it 'serializes register contents as CSV' do
+      step.perform
+
+      expected = "id,name\n1,bozo\n"
+
+      expect(context.register('a')).to eq(expected)
+    end
   end
 end